mcp_server/requirements.txt
# MCP Server Dependencies
mcp>=0.9.0
asyncpg>=0.29.0
python-dotenv>=1.0.1
